requirements.txt
# HTML解析
beautifulsoup4==4.12.3
lxml==5.3.0

# CSS解析
cssutils==2.11.1

# PPTX生成
python-pptx==0.6.23

# 图表截图
playwright==1.45.0

# 图片处理
Pillow==10.4.0

# 工具库
python-dateutil==2.9.0
